Move simulation (libary part)
