+++ bittide-experiments/bittide-experiments.cabal
@@ -79,6 +79,7 @@ library
     aeson,
     aeson-pretty,
     array,
+    bittide,
     bytestring,
     clash-cores,
     clash-lib,
@@ -89,5 +90,38 @@ library
     text
   exposed-modules:
     Bittide.Hitl
+    Bittide.Simulate
+    Bittide.Simulate.ElasticBuffer
+    Bittide.Simulate.Topology
     Bittide.Topology
+  default-language: Haskell2010
+
+test-suite unittests
+  import:           common-options
+  type:             exitcode-stdio-1.0
+  main-is:          unittests.hs
+  ghc-options:      -Wall -Wcompat -threaded -rtsopts "-with-rtsopts=-N -xm20000000" +RTS -xm20000000 -RTS
+  default-language: Haskell2010
+  hs-source-dirs:   tests
+  other-modules:
+    Tests.Bittide.Simulate
+  build-depends:
+      base
+    , bittide
+    , bittide-experiments
+    , clash-prelude
+    , tasty
+    , tasty-hedgehog
+    , tasty-hunit
+
+test-suite doctests
+  type:             exitcode-stdio-1.0
+  hs-source-dirs:   tests
+  main-is:          doctests.hs
+  ghc-options:      -threaded
+  ghc-options:      -Wall -Wcompat -threaded -rtsopts "-with-rtsopts=-xm20000000" +RTS -xm20000000 -RTS
+  build-depends:
+    base,
+    bittide-experiments,
+    doctest-parallel >= 0.3.0.1 && < 0.4,
   default-language: Haskell2010
