+++ elastic-buffer-sim/elastic-buffer-sim.cabal
@@ -89,11 +89,8 @@ library
     text,
     typelits-witnesses
   exposed-modules:
-    Bittide.ClockControl.ElasticBuffer
     Bittide.Domain
     Bittide.Plot
-    Bittide.Simulate
-    Bittide.Simulate.Topology
   default-language: Haskell2010
 
 executable sim
@@ -136,33 +133,3 @@ executable plot
   default-extensions: ImplicitPrelude
   -- enable rtsopts so we can setup memory limits
   ghc-options: -Wall -Wcompat -threaded -rtsopts -with-rtsopts=-xm20000000 +RTS -xm20000000 -RTS
-
-test-suite unittests
-  import:           common-options
-  type:             exitcode-stdio-1.0
-  main-is:          unittests.hs
-  ghc-options:      -Wall -Wcompat -threaded -rtsopts "-with-rtsopts=-N -xm20000000" +RTS -xm20000000 -RTS
-  default-language: Haskell2010
-  hs-source-dirs:   tests
-  other-modules:
-    Tests.Bittide.Simulate
-  build-depends:
-      base
-    , bittide
-    , clash-prelude
-    , elastic-buffer-sim
-    , tasty
-    , tasty-hedgehog
-    , tasty-hunit
-
-test-suite doctests
-  type:             exitcode-stdio-1.0
-  hs-source-dirs:   tests
-  main-is:          doctests.hs
-  ghc-options:      -threaded
-  ghc-options:      -Wall -Wcompat -threaded -rtsopts -with-rtsopts="-xm20000000" +RTS -xm20000000 -RTS
-  build-depends:
-    base,
-    elastic-buffer-sim,
-    doctest-parallel >= 0.3.0.1 && < 0.4,
-  default-language: Haskell2010
